refactor: split concentration issuer mapping
