+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -7,7 +7,7 @@ This ledger tracks the enterprise refactor continuation from
 | --- | --- | --- | --- | --- | --- |
 | RISK-REF-001 | Quality measurement and CI truthfulness | Reviewing | The generated `baseline_report.md` described current state as the initial baseline, mixed source and test hotspots, and the quality-gate guide described already-active gates as report-only. This weakened before/after evidence credibility. | Correct generated terminology and gate posture; preserve commit `3254774` as immutable initial baseline; add generator regression tests next. | `make quality-baseline`; initial baseline verified with `git show 3254774:quality/baseline_report.md` |
 | RISK-REF-002 | Architecture dependency boundaries | Hardened | Risk calculation orchestration and period metric services depended directly on Prometheus `Histogram`, coupling business calculations to observability infrastructure. | Introduced a narrow metric-duration observer port, moved Prometheus metric ownership into `app.observability`, and added an import-linter contract prohibiting Prometheus imports from services. Continue reviewing service-to-infrastructure imports before introducing broader layers. | `tests/unit/test_risk_metric_timing_port.py`; focused risk tests; `make architecture-gate`; `python -m mypy --config-file mypy.ini` |
-| RISK-REF-003 | Service hotspot modularity | Hardened | Current largest source modules included `risk/period_metrics.py`, `rolling_stateful_inputs.py`, `rolling_metric_series.py`, and concentration parsing/orchestration modules. `risk/period_metrics.py` mixed non-benchmark period calculation orchestration with benchmark alignment/error-context handling; `rolling_stateful_inputs.py` mixed dependency selection/reporting-currency resolution with source fetching and parsing; `rolling_metric_series.py` mixed benchmark alignment/rolling beta/tracking-error/information-ratio logic with non-benchmark rolling metrics. | Extracted benchmark-dependent period metric alignment, error handling, and context construction into `risk/benchmark_period_metrics.py`; extracted rolling dependency selection and reporting-currency resolution into `rolling_stateful_dependency_selection.py`; extracted rolling benchmark metric-series calculations into `rolling_benchmark_metric_series.py`. Continue reviewing concentration hotspots. | `tests/unit/test_risk_engine.py`; `tests/unit/test_risk_engine_branch_coverage.py`; `tests/unit/test_risk_mode_adapter_characterization.py`; `tests/integration/test_risk_calculate.py`; rolling focused tests; `python -m mypy --config-file mypy.ini`; touched radon MI/CC checks |
+| RISK-REF-003 | Service hotspot modularity | Hardened | Current largest source modules included `risk/period_metrics.py`, `rolling_stateful_inputs.py`, `rolling_metric_series.py`, and concentration parsing/orchestration modules. `risk/period_metrics.py` mixed non-benchmark period calculation orchestration with benchmark alignment/error-context handling; `rolling_stateful_inputs.py` mixed dependency selection/reporting-currency resolution with source fetching and parsing; `rolling_metric_series.py` mixed benchmark alignment/rolling beta/tracking-error/information-ratio logic with non-benchmark rolling metrics; concentration parsing mixed issuer mapping/enrichment with position parsing and display-name handling. | Extracted benchmark-dependent period metric alignment, error handling, and context construction into `risk/benchmark_period_metrics.py`; extracted rolling dependency selection and reporting-currency resolution into `rolling_stateful_dependency_selection.py`; extracted rolling benchmark metric-series calculations into `rolling_benchmark_metric_series.py`; extracted concentration issuer mapping/enrichment logic into `concentration/issuer_mapping.py`. Continue reviewing remaining concentration parsing/resolver hotspots. | `tests/unit/test_risk_engine.py`; `tests/unit/test_risk_engine_branch_coverage.py`; `tests/unit/test_risk_mode_adapter_characterization.py`; `tests/integration/test_risk_calculate.py`; rolling focused tests; concentration focused tests; `python -m mypy --config-file mypy.ini`; touched radon MI/CC checks |
 | RISK-REF-004 | Security and sensitive-data controls | Hardened | Correlation middleware reflected/logged unbounded caller correlation IDs and preserved malformed or mismatched trace context. Upstream classifiers exposed raw downstream response and transport exception text. Early rejection responses skipped policy headers, and sensitive API responses lacked explicit no-cache/nosniff controls. | Added bounded correlation identity, strict W3C trace-context validation, bounded client-facing upstream messages, consistent enterprise/security response headers, and negative tests. Continue reviewing operational diagnostic redaction. | `tests/unit/test_correlation_middleware.py`; `tests/unit/test_upstream_errors.py`; `tests/unit/test_enterprise_readiness.py`; `quality/security_findings.md`; `make security-audit` |
 | RISK-REF-005 | OpenAPI and error governance | Hardened | Operation metadata is actively gated; the Lotus error envelope previously lacked problem-details-compatible fields. | Added additive RFC 7807/problem-details metadata inside the existing `error` object, preserving `error.code`, `error.message`, `error.correlation_id`, and `error.details` compatibility. Continue regenerating final OpenAPI artifact evidence before PR. | `tests/unit/test_error_response.py`; `tests/integration/test_health.py`; `make openapi-gate`; `make openapi-artifact-gate` |
 | RISK-REF-006 | Observability and operational diagnostics | Hardened | Request middleware emitted an interpolated text log, weakening production parsing and field-level sensitive-data proof. | Replaced it with a bounded structured `request_observation` event and tests proving query strings and bodies are excluded. Continue reviewing operator diagnostics and production threshold evidence. | `tests/unit/test_correlation_middleware.py`; `make observability-contract-validate`; observability tests |
